app/google_app_engine/requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.12.0
google-cloud-storage==1.23.0
google-cloud-logging==1.14.0
google-cloud-error-reporting==0.33.0
gunicorn==19.9.0
six>=1.11.0
